requirements.txt
discord
sqlite3
python-dotenv